Add files via upload
We have a new tool that can detect and recognize faces. . It is an easy-to-use Python package that can do face detection and recognition tasks. The model is also very small in size, making it a great choice for mobile or web applications. You can also use this to detect faces in images from video streams. 

We are very excited to be able to share it with you. You can find our code here: